[chores] Misc improvements
- renamed metrics_collection to metric_collection
- reduced verbosity
- isolated code meant to be used in admin_theme.admin
- avoided used of term "disclaimer"
- moved most of the code in class methods of models.OpenwispVersion
