+++ openwisp_utils/metric_collection/static/admin/js/metric-collection-consent.js
