+++ openwisp_utils/metric_collection/tasks.py
@@ -1,0 +1,9 @@
+from celery import shared_task
+
+from ..tasks import OpenwispCeleryTask
+from .models import OpenwispVersion
+
+
+@shared_task(base=OpenwispCeleryTask)
+def send_usage_metrics(category='Heartbeat'):
+    OpenwispVersion.send_usage_metrics(category)
