+++ runtests.py
@@ -13,5 +13,5 @@ if __name__ == "__main__":
     args = sys.argv
     args.insert(1, "test")
     args.insert(2, "test_project")
-    args.insert(3, "openwisp_utils.metrics_collection")
+    args.insert(3, "openwisp_utils.metric_collection")
     execute_from_command_line(args)
